+++ packages/stargate-ui-core/src/Card/Card.jsx
@@ -21,7 +21,7 @@ const styles = (theme) => {
   );
 
   const card = {
-    display: 'flex',
+    display: 'block',
     maxWidth: 'max-content',
     cursor: ({ clickable }) => (
       clickable && 'pointer'
@@ -135,7 +135,8 @@ Card.propTypes = {
     'success',
     'warning',
     'info',
-    'error'
+    'error',
+    false,
   ]),
   /**
    * Trigger when element is clicked.
@@ -145,7 +146,7 @@ Card.propTypes = {
 };
 
 /**
- * Add @property {object} factoryProps made available properties information
+ * Add @property {object} defaultProps made available properties information
  * for Props in the Storybook but do not use as major define for default properties.
  */
 Card.defaultProps = {