+++ packages/stargate-ui-core/src/Svg/Svg.js
@@ -1,26 +1,38 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
+import { useStyles } from '@pontte/stargate-ui-styles';
 
 import Factory from '../Factory';
 
-const Svg = (props) => {
+const styles = () => {
+  const svg = {
+    userSelect: 'none',
+    pointerEvents: 'none',
+  };
+
+  return { svg };
+};
+
+const Svg = React.forwardRef((props, ref) => {
   const {
     className: inheritedClassName,
     ...factoryProps
   } = props;
 
-  const className = clsx(inheritedClassName);
+  const [{ svg }] = useStyles(styles);
+  const className = clsx(svg, inheritedClassName);
 
   return (
     <Factory
+      ref={ref}
       element="span"
-      className={className}
       aria-hidden
       {...factoryProps}
+      className={className}
     />
   );
-};
+});
 
 Svg.propTypes = {
   children: PropTypes.element,